refactor to allow packaging of oracledb_exporter
